+++ spec/support/character.rb
@@ -2,5 +2,6 @@ class Character < ActiveRecord::Base
   # string :name
   # string :description
   # integer :web_comic_id
+
   belongs_to :web_comic
 end