+++ spec/support/game.rb
@@ -3,5 +3,3 @@ class Game < ActiveRecord::Base
   # string :title
   # text :description
 end
-
-class GameFail < Game; end