+++ spec/support/not_there.rb
@@ -1,0 +1,3 @@
+require 'support/ar_stand_in'
+
+class NotThere < ARStandIn; end