+++ spec/support/web_comic_with_searchable.rb
@@ -1,0 +1,6 @@
+require 'textacular/searchable'
+require 'support/web_comic'
+
+class WebComicWithSearchable < WebComic
+  extend Searchable
+end