+++ spec/textacular/migration_generator_spec.rb
@@ -1,68 +1,67 @@
-require 'spec_helper'
-
-class MigrationGeneratorTest < Test::Unit::TestCase
-  context ".stream_output" do
+RSpec.describe Textacular::MigrationGenerator do
+  describe ".stream_output" do
     context "when Rails is not defined" do
-      setup do
-        @generator = Textacular::MigrationGenerator.new('filename', 'content')
+      subject do
+        Textacular::MigrationGenerator.new('filename', 'content')
       end
 
-      should "point to stdout" do
-        @output_stream = nil
+      it "points to STDOUT" do
+        output_stream = nil
 
-        @generator.stream_output do |io|
-          @output_stream = io
+        subject.stream_output do |io|
+          output_stream = io
         end
 
-        assert_equal(@output_stream, $stdout)
+        expect(output_stream).to eq(STDOUT)
       end
     end
 
-    context "When Rails IS defined" do
-      setup do
+    context "when Rails is defined" do
+      before do
         module ::Rails
           # Stub this out, sort of.
           def self.root
             File.join('.', 'fake_rails')
           end
         end
-
-        @now = Time.now
-        @filename = 'filename'
-        @generator = Textacular::MigrationGenerator.new(@filename, 'content')
       end
 
-      teardown do
+      after do
         Object.send(:remove_const, :Rails)
         FileUtils.rm_rf(File.join('.', 'fake_rails'))
       end
 
-      should "point to a properly named migration file" do
-        expected_file_name = "./fake_rails/db/migrate/#{@now.strftime('%Y%m%d%H%M%S')}_#{@filename}.rb"
+      let(:now) do
+        Time.now
+      end
 
-        @output_stream = nil
+      subject do
+        Textacular::MigrationGenerator.new('file_name', 'content')
+      end
 
-        @generator.stream_output(@now) do |io|
-          @output_stream = io
+      it "points to a properly names migration file" do
+        expected_file_name = "./fake_rails/db/migrate/#{now.strftime('%Y%m%d%H%M%S')}_file_name.rb"
+
+        output_stream = nil
+
+        subject.stream_output(now) do |io|
+          output_stream = io
         end
 
-        assert_equal(expected_file_name, @output_stream.path)
+        expect(output_stream.path).to eq(expected_file_name)
       end
     end
-  end
 
-  context "the content of the migration" do
-    setup do
-      @migration_content = "content\n" # newline automatically added
-      @generator = Textacular::MigrationGenerator.new('filename', @migration_content)
-      @output = StringIO.new
-      @generator.instance_variable_set(:@output_stream, @output)
-    end
+    it "generates the right SQL" do
+      content = "content\n" #newline automatically added
+      output = StringIO.new
 
-    should "generate the right sql" do
-      @generator.generate_migration
+      generator = Textacular::MigrationGenerator.new('file_name', content)
+      generator.instance_variable_set(:@output_stream, output)
 
-      assert_equal(@migration_content, @output.string)
+      generator.generate_migration
+
+      expect(output.string).to eq(content)
     end
   end
 end