+++ spec/textacular/searchable_spec.rb
@@ -1,102 +1,183 @@
-require 'spec_helper'
-require 'textacular/searchable'
+require 'support/web_comic_with_searchable'
+require 'support/web_comic_with_searchable_name'
+require 'support/web_comic_with_searchable_name_and_author'
+require 'support/character'
 
-class SearchableTest < Test::Unit::TestCase
+RSpec.describe "Searchable" do
   context "when extending an ActiveRecord::Base subclass" do
     context "with no parameters" do
-      setup do
-        @qcont = WebComicWithSearchable.create :name => "Questionable Content", :author => "Jeph Jaques"
-        @jhony = WebComicWithSearchable.create :name => "Johnny Wander", :author => "Ananth & Yuko"
-        @ddeeg = WebComicWithSearchable.create :name => "Dominic Deegan", :author => "Mookie"
-        @penny = WebComicWithSearchable.create :name => "Penny Arcade", :author => "Tycho & Gabe"
-        @null  = WebComicWithSearchable.create :author => 'Foo'
+      let!(:questionable_content) do
+        WebComicWithSearchable.create(
+          name: 'Questionable Content',
+          author: 'Jeph Jaques',
+        )
       end
 
-      teardown do
-        WebComicWithSearchable.delete_all
+      let!(:johnny_wander) do
+        WebComicWithSearchable.create(
+          name: 'Johnny Wander',
+          author: 'Ananth & Yuko',
+        )
       end
 
-      should "search across all columns" do
-        assert_equal [@penny], WebComicWithSearchable.advanced_search("Penny")
-        assert_equal [@ddeeg], WebComicWithSearchable.advanced_search("Dominic")
+      let!(:dominic_deegan) do
+        WebComicWithSearchable.create(
+          name: 'Dominic Deegan',
+          author: 'Mookie',
+        )
       end
 
-      should "still rank with NULL columns" do
+      let!(:penny_arcade) do
+        WebComicWithSearchable.create(
+          name: 'Penny Arcade',
+          author: 'Tycho & Gabe',
+        )
+      end
+
+      let!(:null) do
+        WebComicWithSearchable.create(
+          author: 'Foo',
+        )
+      end
+
+      it "searches across all columns" do
+        expect(
+          WebComicWithSearchable.advanced_search("Penny")
+        ).to eq([penny_arcade])
+        expect(
+          WebComicWithSearchable.advanced_search("Dominic")
+        ).to eq([dominic_deegan])
+      end
+
+      it "ranks results, egen with NULL columns" do
         comic = WebComicWithSearchable.basic_search('Foo').first
         rank = comic.attributes.find { |key, value| key.to_s =~ /\Arank\d+\z/ }.last
 
-        assert rank
+        expect(rank).to be_present
       end
     end
 
-    context "with one column as parameter" do
-      setup do
-        @qcont = WebComicWithSearchableName.create :name => "Questionable Content", :author => "Jeph Jaques"
-        @jhony = WebComicWithSearchableName.create :name => "Johnny Wander", :author => "Ananth & Yuko"
-        @ddeeg = WebComicWithSearchableName.create :name => "Dominic Deegan", :author => "Mookie"
-        @penny = WebComicWithSearchableName.create :name => "Penny Arcade", :author => "Tycho & Gabe"
+    context "with one column as a parameter" do
+      let!(:questionable_content) do
+        WebComicWithSearchableName.create(
+          name: 'Questionable Content',
+          author: 'Jeph Jaques',
+        )
       end
 
-      teardown do
-        WebComicWithSearchableName.delete_all
+      let!(:johnny_wander) do
+        WebComicWithSearchableName.create(
+          name: 'Johnny Wander',
+          author: 'Ananth & Yuko',
+        )
       end
 
-      should "only search across the given column" do
-        assert_equal [@penny], WebComicWithSearchableName.advanced_search("Penny")
-        assert_empty WebComicWithSearchableName.advanced_search("Tycho")
+      let!(:dominic_deegan) do
+        WebComicWithSearchableName.create(
+          name: 'Dominic Deegan',
+          author: 'Mookie',
+        )
       end
 
-      ["hello \\", "tebow!" , "food &"].each do |search_term|
-        should "be fine with searching for crazy character #{search_term} with plain search" do
-          # Uses plainto_tsquery
-          assert_equal [], WebComicWithSearchableName.basic_search(search_term)
-        end
+      let!(:penny_arcade) do
+        WebComicWithSearchableName.create(
+          name: 'Penny Arcade',
+          author: 'Tycho & Gabe',
+        )
+      end
 
-        should "be not fine with searching for crazy character #{search_term} with advanced search" do
-          # Uses to_tsquery
-          assert_raise(ActiveRecord::StatementInvalid) do
-            WebComicWithSearchableName.advanced_search(search_term).first
+      it "only searches across the given column" do
+        expect(WebComicWithSearchableName.advanced_search("Penny")).to eq([penny_arcade])
+
+        expect(WebComicWithSearchableName.advanced_search("Tycho")).to be_empty
+      end
+
+      describe "basic search" do # Uses plainto_tsquery
+        ["hello \\", "tebow!" , "food &"].each do |search_term|
+          it "works with interesting term \"#{search_term}\"" do
+            expect(WebComicWithSearchableName.basic_search(search_term)).to be_empty
           end
         end
       end
 
-      should "fuzzy search stuff" do
-        assert_equal [@qcont], WebComicWithSearchableName.fuzzy_search('Questio')
-      end
-
-      should "define :searchable_columns as private" do
-        assert_raise(NoMethodError) { WebComicWithSearchableName.searchable_columns }
-        begin
-          WebComicWithSearchableName.searchable_columns
-        rescue NoMethodError => error
-          assert_match /private method/, error.message
+      describe "advanced_search" do # Uses to_tsquery
+        ["hello \\", "tebow!" , "food &"].each do |search_term|
+          it "fails with interesting term \"#{search_term}\"" do
+            expect {
+              WebComicWithSearchableName.advanced_search(search_term).first
+            }.to raise_error(ActiveRecord::StatementInvalid)
+          end
         end
       end
 
-      should "define #indexable_columns which returns a write-proof Enumerable" do
-        assert_equal(Enumerator, WebComicWithSearchableName.indexable_columns.class)
-        assert_raise(NoMethodError) { WebComicWithSearchableName.indexable_columns[0] = 'foo' }
+      it "does fuzzy searching" do
+        expect(
+          WebComicWithSearchableName.fuzzy_search('Questio')
+        ).to eq([questionable_content])
+      end
+
+      it "defines :searchable_columns as private" do
+        expect { WebComicWithSearchableName.searchable_columns }.to raise_error(NoMethodError)
+
+        begin
+          WebComicWithSearchableName.searchable_columns
+        rescue NoMethodError => error
+          expect(error.message).to match(/private method/)
+        end
+      end
+
+      it "defines #indexable_columns which returns a write-proof Enumerable" do
+        expect(WebComicWithSearchableName.indexable_columns).to be_an(Enumerator)
+
+        expect {
+          WebComicWithSearchableName.indexable_columns[0] = 'foo'
+        }.to raise_error(NoMethodError)
       end
     end
 
     context "with two columns as parameters" do
-      setup do
-        @qcont = WebComicWithSearchableNameAndAuthor.create :name => "Questionable Content", :author => "Jeph Jaques"
-        @jhony = WebComicWithSearchableNameAndAuthor.create :name => "Johnny Wander", :author => "Ananth & Yuko"
-        @ddeeg = WebComicWithSearchableNameAndAuthor.create :name => "Dominic Deegan", :author => "Mookie"
-        @penny = WebComicWithSearchableNameAndAuthor.create :name => "Penny Arcade", :author => "Tycho & Gabe"
+      let!(:questionable_content) do
+        WebComicWithSearchableNameAndAuthor.create(
+          name: 'Questionable Content',
+          author: 'Jeph Jaques',
+        )
       end
 
-      teardown do
-        WebComicWithSearchableNameAndAuthor.delete_all
+      let!(:johnny_wander) do
+        WebComicWithSearchableNameAndAuthor.create(
+          name: 'Johnny Wander',
+          author: 'Ananth & Yuko',
+        )
       end
 
-      should "only search across the given column" do
-        assert_equal [@penny], WebComicWithSearchableNameAndAuthor.advanced_search("Penny")
-        assert_equal [@penny], WebComicWithSearchableNameAndAuthor.advanced_search("Tycho")
+      let!(:dominic_deegan) do
+        WebComicWithSearchableNameAndAuthor.create(
+          name: 'Dominic Deegan',
+          author: 'Mookie',
+        )
       end
-      should "allow includes" do
-        assert_equal [@penny], WebComicWithSearchableNameAndAuthor.includes(:characters).advanced_search("Penny")
+
+      let!(:penny_arcade) do
+        WebComicWithSearchableNameAndAuthor.create(
+          name: 'Penny Arcade',
+          author: 'Tycho & Gabe',
+        )
+      end
+
+      it "only searches across the given columns" do
+        expect(
+          WebComicWithSearchableNameAndAuthor.advanced_search("Penny")
+        ).to eq([penny_arcade])
+
+        expect(
+          WebComicWithSearchableNameAndAuthor.advanced_search("Tycho")
+        ).to eq([penny_arcade])
+      end
+
+      it "allows includes" do
+        expect(
+          WebComicWithSearchableNameAndAuthor.includes(:characters).advanced_search("Penny")
+        ).to eq([penny_arcade])
       end
     end
   end